requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0

# Data
datasets>=2.18.0

# LLM Judge
google-generativeai>=0.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0

# Notebook
jupyter>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
